Convert login and registration to use SRP.
The Secure Remote Password authentication protocol allows a client to
provide proof of having the correct password without ever transmitting
the password to the server.

The SRP verification secret and verification salt are put through the
AES cipher in the cipher block chaining (CBC) mode with a unique
initialization vector for each user using an application-specific secret key
to increase the computation needed to brute-force the data at rest if
it's ever obtained by a third party.